Rework the structure of the classes config
This should provide for more flexibility, in particular when wanting to
override classes for a specific input type in a specific state.

However, for now it means we miss the ability to configure classes for
inputs when rendered in an input group. This will be added back in a
future commit.