Pass false to not render a label
- If text is present, we'll use that text as the label content.

- If text is nil or blank, we'll fall back to whatever
ActionView::Helpers::FormBuilder#label does.

- If text is false we'll not render a label.

Originally I was going to use `nil` to mean "don't render a label", but
that is not how ActionView::Helpers::FormBuilder#label works. When given
`nil` - or any non-present object - it will use its default behavior.

So for the sake of preserving principle of least surprise we'll use
`false` to not render a label. It seems more explicit than `nil` and it
doesn't go against what ActionView does.